# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Encryptr Server API",
    "description": "Encrypted AI conversation backend. Message content is always client-side AEAD ciphertext; the server stores ciphertext, nonce, and auth tag but can never decrypt it. The `/chat/stream` WebSocket provides end-to-end encrypted streaming inference.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List messages in a conversation (newest first, paginated).",
        "operationId": "list",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "query",
            "description": "Conversation whose messages to return.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, clamped server-side to 1..=200 (default 50).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Return messages created strictly before this timestamp (pagination).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Store a client-encrypted message. The server never sees plaintext.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/messages/{id}": {
      "delete": {
        "tags": [
          "messages"
        ],
        "summary": "Delete one of the caller's messages.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Message id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The error envelope returned by every endpoint on failure. Mirrors\n`crate::error::AppError`'s HTTP body (defined here so it can carry a\n`ToSchema` derive without leaking into the error type).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `validation_error`, `unauthorized`,\n`not_found`, `conflict`, `rate_limited`, `provider_error`,\n`internal_error`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, safe-to-display message."
          }
        }
      },
...
      "CreateMessageRequest": {
        "type": "object",
        "required": [
          "conversation_id",
          "role",
          "ciphertext",
          "nonce"
        ],
        "properties": {
          "algorithm": {
            "type": [
              "string",
              "null"
            ]
          },
          "ciphertext": {
            "type": "string",
            "description": "Base64-encoded AEAD ciphertext (tag appended), produced client-side."
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "key_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "nonce": {
            "type": "string",
            "description": "Base64-encoded 96-bit nonce."
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "role",
          "ciphertext",
          "nonce",
          "algorithm",
          "created_at"
        ],
        "properties": {
          "algorithm": {
            "type": "string"
          },
          "ciphertext": {
            "type": "string"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "nonce": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /login`."
      }
    }
  },
...
}